requirements-main.txt
aiofile>=3.3.3
aiofiles>=0.6.0
aiohttp>=3.7.3
asyncio>=3.4.3
docker-py>=1.10.6
django>=3.1.5
flask>=1.1.2
jinja2>=2.11.2
PyYAML>=5.3.1
requests>=2.25.1
xlsxwriter>=1.3.7
yajl>=0.3.5
selenium>=3.141.0
Werkzeug>=1.0.1
cryptography>=2.8
